refactor(useFavicon): Refactor code, add missing documentation, improve example
This commit adds remaining things that are required for adding this hook to the library. Missing
documentation is added, example is improved, ESLint warnings fixed and import style fixed.

re #33
