+++ src/__docs__/migrating-from-react-use.story.mdx
@@ -1,4 +1,4 @@
-import { Meta } from '@storybook/addon-docs';
+import { Meta } from '@storybook/addon-docs'
 
 <Meta title="Migrating from react-use" />
 
@@ -434,7 +434,7 @@ Not implemented yet
 
 #### useFavicon
 
-Not implemented yet
+Implemented as [useFavicon](/docs/side-effect-usefavicon--example)
 
 #### useLocalStorage
